django/requirements/base.txt
Django>=2.1.5
python-dotenv>=0.7.1
django-rest-framework>=0.1.0
coreapi>=2.3.3
PyYAML>=3.13
